# [Performance, CPU] Rewriting OpenMP pragmas into parallel_for