--- a/src/array/cpu/array_pack.cc
+++ b/src/array/cpu/array_pack.cc
@@ -31,9 +31,11 @@ std::pair<NDArray, IdArray> ConcatSlices(NDArray array, IdArray lengths) {
   NDArray concat = NDArray::Empty({total_length}, array->dtype, array->ctx);
   DType *concat_data = static_cast<DType *>(concat->data);
 
-  parallel_for(0, rows, [=](size_t i) {
-    for (int64_t j = 0; j < length_data[i]; ++j)
-      concat_data[offsets_data[i] + j] = array_data[i * stride + j];
+  parallel_for(0, rows, [=](size_t b, size_t e) {
+    for (auto i = b; i < e; ++i) {
+      for (int64_t j = 0; j < length_data[i]; ++j)
+        concat_data[offsets_data[i] + j] = array_data[i * stride + j];
+    }
   });
 
   return std::make_pair(concat, offsets);
@@ -57,14 +59,16 @@ std::tuple<NDArray, IdArray, IdArray> Pack(NDArray array, DType pad_value) {
 
   IdArray length = NewIdArray(rows, array->ctx);
   int64_t *length_data = static_cast<int64_t *>(length->data);
-  parallel_for(0, rows, [=](size_t i) {
-    int64_t j;
-    for (j = 0; j < cols; ++j) {
-      const DType val = array_data[i * cols + j];
-      if (val == pad_value)
-        break;
+  parallel_for(0, rows, [=](size_t b, size_t e) {
+    for (auto i = b; i < e; ++i) {
+      int64_t j;
+      for (j = 0; j < cols; ++j) {
+        const DType val = array_data[i * cols + j];
+        if (val == pad_value)
+          break;
+      }
+      length_data[i] = j;
     }
-    length_data[i] = j;
   });
 
   auto ret = ConcatSlices<XPU, DType, int64_t>(array, length);