--- a/src/array/cpu/csr_sum.cc
+++ b/src/array/cpu/csr_sum.cc
@@ -27,13 +27,15 @@ void CountNNZPerRow(
     int64_t M) {
   int64_t n = A_indptr.size();
 
-  runtime::parallel_for(0, M, [=](size_t i) {
-    phmap::flat_hash_set<IdType> set;
-    for (int64_t k = 0; k < n; ++k) {
-      for (IdType u = A_indptr[k][i]; u < A_indptr[k][i + 1]; ++u)
-        set.insert(A_indices[k][u]);
+  runtime::parallel_for(0, M, [=](size_t b, size_t e) {
+    for (size_t i = b; i < e; ++i) {
+      phmap::flat_hash_set<IdType> set;
+      for (int64_t k = 0; k < n; ++k) {
+        for (IdType u = A_indptr[k][i]; u < A_indptr[k][i + 1]; ++u)
+          set.insert(A_indices[k][u]);
+      }
+      C_indptr_data[i] = set.size();
     }
-    C_indptr_data[i] = set.size();
   });
 }
 
@@ -61,20 +63,22 @@ void ComputeIndicesAndData(
     DType* C_weights_data,
     int64_t M) {
   int64_t n = A_indptr.size();
-  runtime::parallel_for(0, M, [=](size_t i) {
-    phmap::flat_hash_map<IdType, DType> map;
-    for (int64_t k = 0; k < n; ++k) {
-      for (IdType u = A_indptr[k][i]; u < A_indptr[k][i + 1]; ++u) {
-        IdType kA = A_indices[k][u];
-        DType vA = A_data[k][A_eids[k] ? A_eids[k][u] : u];
-        map[kA] += vA;
+  runtime::parallel_for(0, M, [=](size_t b, size_t e) {
+    for (auto i = b; i < e; ++i) {
+      phmap::flat_hash_map<IdType, DType> map;
+      for (int64_t k = 0; k < n; ++k) {
+        for (IdType u = A_indptr[k][i]; u < A_indptr[k][i + 1]; ++u) {
+          IdType kA = A_indices[k][u];
+          DType vA = A_data[k][A_eids[k] ? A_eids[k][u] : u];
+          map[kA] += vA;
+        }
       }
-    }
-    IdType j = C_indptr_data[i];
-    for (auto it : map) {
-      C_indices_data[j] = it.first;
-      C_weights_data[j] = it.second;
-      ++j;
+      IdType j = C_indptr_data[i];
+      for (auto it : map) {
+        C_indices_data[j] = it.first;
+        C_weights_data[j] = it.second;
+        ++j;
+      }
     }
   });
 }