The PR replaces uses of OpenMP `parallel for` pragma with `parallel_for` primitive wherever it was possible (no `reduction`, `atomic`, `critical` or `barrier` clauses). 

This PR continues and includes work introduced by #3023 and #3027. The mentioned PRs contain initial benchmarking and information about performance regressions. This PR includes commits from earlier PRs and replaces additional pragmas. The earlier PR are going to be closed.