@tpatejko 2021-07-28
@dgl-bot run c5.9xlarge kernel,api dmlc/master