@BarclayII 2021-08-03
The largest speedup seems to be readout functions with 25%.

Here is the list of the benchmarks slowed down by 40%+.

| test_name | params | unit | number_master | machine | number_3171 | ratio |
|-|-|-|-|-|-|-|
| api.bench_edge_ids.track_time | 'cora', 'coo', 0.01, False | s | 0.000208 | c5.9xlarge-cpu | 0.000396 | 0.902062 |
| api.bench_in_degrees.track_time | 'cora', 'coo', 0.01 | s | 5.22E-05 | c5.9xlarge-cpu | 9.77E-05 | 0.87219 |
| api.bench_knn_graph.track_time | 10000, 32, 64, 'nn-descent' | s | 0.692811 | c5.9xlarge-cpu | 1.076581 | 0.553932 |
| api.bench_edge_ids.track_time | 'cora', 'coo', 0.1, False | s | 0.000462 | c5.9xlarge-cpu | 0.000689 | 0.491037 |
| api.bench_knn_graph.track_time | 10000, 32, 8, 'nn-descent' | s | 0.072226 | c5.9xlarge-cpu | 0.104481 | 0.446579 |
| api.bench_builtin_apply_edges.track_time | 'cora', 'coo', 512, 'u+v' | s | 0.0013 | c5.9xlarge-cpu | 0.001876 | 0.442883 |
| api.bench_knn_graph.track_time | 1000, 4, 8, 'nn-descent' | s | 0.007045 | c5.9xlarge-cpu | 0.010088 | 0.431965 |
| api.bench_knn_graph.track_time | 10000, 4, 8, 'nn-descent' | s | 0.083871 | c5.9xlarge-cpu | 0.118583 | 0.413867 |