@BarclayII 2021-08-16
The benchmark result goes as follows, with some of the `update_all` execution on CSC gets 30% slower (mostly `u_mul_e` message function and `sum` reduce function on small feature sizes)

  | test_name | params | unit | number_3171 | machine | number_master | ratio
-- | -- | -- | -- | -- | -- | -- | --
8 | api.bench_builtin_apply_edges.track_time | 'cora', 'coo', 128, 'u->e' | s | 0.000407 | c5.9xlarge-cpu | 0.000195 | 1.086563
197 | api.bench_homograph_scipy_construction.track_time | 'small', 'coo' | s | 0.000137 | c5.9xlarge-cpu | 6.81E-05 | 1.009174
9 | api.bench_builtin_apply_edges.track_time | 'cora', 'coo', 128, 'u+v' | s | 0.000473 | c5.9xlarge-cpu | 0.000247 | 0.912796
16 | api.bench_builtin_apply_edges.track_time | 'cora', 'csr', 512, 'u->e' | s | 0.001269 | c5.9xlarge-cpu | 0.000767 | 0.655484
374 | api.bench_random_walk.track_time | 'livejournal', 1000, 20, '_node2vec' | s | 0.001473 | c5.9xlarge-cpu | 0.000946 | 0.557238
69 | api.bench_builtin_update_all_csc.track_time | 'ogbn-arxiv', 'csc', 4, 'u_mul_e', 'sum' | s | 0.002903 | c5.9xlarge-cpu | 0.002016 | 0.440368
237 | api.bench_knn_graph.track_time | 1000, 4, 8, 'bruteforce-blas' | s | 0.00328 | c5.9xlarge-cpu | 0.002344 | 0.398937
11 | api.bench_builtin_apply_edges.track_time | 'cora', 'coo', 512, 'u+v' | s | 0.001333 | c5.9xlarge-cpu | 0.000972 | 0.372171
427 | api.bench_readout.track_time | 256, 16, 'min', 'edge' | s | 0.000238 | c5.9xlarge-cpu | 0.000177 | 0.344443
88 | api.bench_builtin_update_all_csc.track_time | 'reddit', 'csc', 4, 'u_mul_e', 'mean' | s | 0.39102 | c5.9xlarge-cpu | 0.291074 | 0.343371
87 | api.bench_builtin_update_all_csc.track_time | 'reddit', 'csc', 4, 'u_mul_e', 'sum' | s | 0.389416 | c5.9xlarge-cpu | 0.289952 | 0.343036
106 | api.bench_builtin_update_all_csc.track_time | 'ogbn-proteins', 'csc', 4, 'u_mul_e', 'mean' | s | 0.269016 | c5.9xlarge-cpu | 0.200894 | 0.339094
70 | api.bench_builtin_update_all_csc.track_time | 'ogbn-arxiv', 'csc', 4, 'u_mul_e', 'mean' | s | 0.00373 | c5.9xlarge-cpu | 0.0028 | 0.331844
105 | api.bench_builtin_update_all_csc.track_time | 'ogbn-proteins', 'csc', 4, 'u_mul_e', 'sum' | s | 0.268104 | c5.9xlarge-cpu | 0.2016 | 0.329882
229 | api.bench_in_subgraph.track_time | 'reddit', 'csc', 5000 | s | 0.023506 | c5.9xlarge-cpu | 0.017932 | 0.310838
125 | api.bench_edge_ids.track_time | 'cora', 'csr', 0.01, False | s | 0.000227 | c5.9xlarge-cpu | 0.000174 | 0.303494