@tpatejko 2021-08-19
@BarclayII Could you trigger performance tests?

I changed the `parallel_for` implementation in order to limit number of call to task lambdas. For the performance tests with small data sizes repeated function calls in the loop could result in additional overhead and performance regression.

In this implementation, `parallel_for` requires a lambda that takes two parameters, beginning and end of the range. There's also an additional for-loop in the lambda's body.