@tpatejko 2021-08-23
@BarclayII I see that there are still some regressions. I collected the results for cases with 20% slowdown.
I also executed the performance tests on our internal CI. Most of the cases are confirmed, with exception for some `readout` cases, `sample_neighbors` and `udf_update_all` cases.

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:///C:/Users/tpatejko/AppData/Local/Temp/msohtmlclip1/01/clip.htm">
<link rel=File-List
href="file:///C:/Users/tpatejko/AppData/Local/Temp/msohtmlclip1/01/clip_filelist.xml">
<style>
<!--table
	{mso-displayed-decimal-separator:"\.";
	mso-displayed-thousand-separator:"\,";}
@page
	{margin:.75in .7in .75in .7in;
	mso-header-margin:.3in;
	mso-footer-margin:.3in;}
tr
	{mso-height-source:auto;}
col
	{mso-width-source:auto;}
br
	{mso-data-placement:same-cell;}
td
	{padding-top:1px;
	padding-right:1px;
	padding-left:1px;
	mso-ignore:padding;
	color:black;
	font-size:11.0pt;
	font-weight:400;
	font-style:normal;
	text-decoration:none;
	font-family:Calibri, sans-serif;
	mso-font-charset:0;
	mso-number-format:General;
	text-align:general;
	vertical-align:bottom;
	border:none;
	mso-background-source:auto;
	mso-pattern:auto;
	mso-protection:locked visible;
	white-space:nowrap;
	mso-rotate:0;}
-->
</style>
</head>

<body link="#0563C1" vlink="#954F72">



  |   |   |   | Github |   |   | Internal |   |  
-- | -- | -- | -- | -- | -- | -- | -- | -- | --
  |   |   |   | Master | PR | Ratio | Master | PR | Ratio
69 | api.bench_builtin_update_all_csc.track_time | 'ogbn-arxiv', 'csc', 4, 'u_mul_e', 'sum' | s | 0.002142 | 0.002978 | 39.00667 | 0.002098 | 0.00249 | 18.6674
70 | api.bench_builtin_update_all_csc.track_time | 'ogbn-arxiv', 'csc', 4, 'u_mul_e', 'mean' | s | 0.002872 | 0.003724 | 29.65746 | 0.002831 | 0.003485 | 23.11899
87 | api.bench_builtin_update_all_csc.track_time | 'reddit', 'csc', 4, 'u_mul_e', 'sum' | s | 0.288187 | 0.389302 | 35.08679 | 0.251854 | 0.300917 | 19.48072
88 | api.bench_builtin_update_all_csc.track_time | 'reddit', 'csc', 4, 'u_mul_e', 'mean' | s | 0.290025 | 0.394061 | 35.87135 | 0.25385 | 0.302694 | 19.24125
105 | api.bench_builtin_update_all_csc.track_time | 'ogbn-proteins', 'csc', 4, 'u_mul_e', 'sum' | s | 0.203356 | 0.264293 | 29.96567 | 0.191394 | 0.220862 | 15.39679
106 | api.bench_builtin_update_all_csc.track_time | 'ogbn-proteins', 'csc', 4, 'u_mul_e', 'mean' | s | 0.203602 | 0.266956 | 31.11676 | 0.191018 | 0.220678 | 15.52727
129 | api.bench_edge_ids.track_time | 'cora', 'csc', 0.01, False | s | 0.000189 | 0.000265 | 39.99919 | 0.000252 | 0.000298 | 18.0397
262 | api.bench_knn_graph.track_time | 10000, 4, 8, 'bruteforce' | s | 0.040429 | 0.050585 | 25.12037 | 0.050068 | 0.055993 | 11.8345
431 | api.bench_readout.track_time | 256, 128, 'sum', 'edge' | s | 0.000343 | 0.000476 | 38.87093 | 0.000243 | 0.00037 | 52.14224
437 | api.bench_readout.track_time | 256, 128, 'mean', 'edge' | s | 0.000397 | 0.000526 | 32.47723 | 0.000335 | 0.000376 | 12.08825
439 | api.bench_readout.track_time | 256, 512, 'sum', 'edge' | s | 0.001546 | 0.001959 | 26.69503 | 0.001155 | 0.00117 | 1.335369
445 | api.bench_readout.track_time | 256, 512, 'mean', 'edge' | s | 0.001618 | 0.002037 | 25.84721 | 0.001332 | 0.001304 | -2.06112
514 | api.bench_sample_neighbors.track_time | 'reddit', 'csc', 20000, 20 | s | 0.004243 | 0.005482 | 29.22649 | 0.003689 | 0.003499 | -5.14895
574 | api.bench_udf_update_all.track_time | 'pubmed', 'coo', 8, 'copy_u', 'sum' | s | 0.134991 | 0.163856 | 21.38245 | 0.150958 | 0.148873 | -1.38128



</body>

</html>


I'm not quite sure how to proceed with the PR. The implementation in the PR of the `parallel_for` primitive is similar to the one in Pytorch: the interface and the implementation are changed in order to reduce the number of function calls to the task per thread.

We also executed training with the changes in the PR and no performance regressions were encountered:
<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:///C:/Users/tpatejko/AppData/Local/Temp/msohtmlclip1/01/clip.htm">
<link rel=File-List
href="file:///C:/Users/tpatejko/AppData/Local/Temp/msohtmlclip1/01/clip_filelist.xml">
<style>
<!--table
	{mso-displayed-decimal-separator:"\.";
	mso-displayed-thousand-separator:"\,";}
@page
	{margin:.75in .7in .75in .7in;
	mso-header-margin:.3in;
	mso-footer-margin:.3in;}
tr
	{mso-height-source:auto;}
col
	{mso-width-source:auto;}
br
	{mso-data-placement:same-cell;}
td
	{padding-top:1px;
	padding-right:1px;
	padding-left:1px;
	mso-ignore:padding;
	color:black;
	font-size:11.0pt;
	font-weight:400;
	font-style:normal;
	text-decoration:none;
	font-family:Calibri, sans-serif;
	mso-font-charset:0;
	mso-number-format:General;
	text-align:general;
	vertical-align:bottom;
	border:none;
	mso-background-source:auto;
	mso-pattern:auto;
	mso-protection:locked visible;
	white-space:nowrap;
	mso-rotate:0;}
.xl63
	{font-weight:700;
	font-family:"Segoe UI", sans-serif;
	mso-font-charset:0;
	text-align:center;
	vertical-align:middle;
	border:.5pt solid black;
	white-space:normal;}
.xl64
	{font-family:"Segoe UI", sans-serif;
	mso-font-charset:0;
	vertical-align:middle;
	border:.5pt solid black;
	white-space:normal;}
-->
</style>
</head>

<body link="#0563C1" vlink="#954F72">

<div style='box-sizing: border-box;font-variant-ligatures: normal;font-variant-caps: normal;
orphans: 2;text-align:start;widows: 2'>



model | dataset | master | PR | speedup
-- | -- | -- | -- | --
gat_ns | ogbn-products | 208.451482 | 208.422 | 1
gat_ns | reddit | 73.361161 | 65.3351 | 1.12
rgcn_hetero_ns | am | 5.270454 | 2.58049 | 2.04
rgcn_hetero_ns | ogbn-mag | 142.622971 | 65.5553 | 2.18
rgcn_homogeneous_ns | am | 1.532482 | 1.48914 | 1.03
rgcn_homogeneous_ns | ogbn-mag | 268.887767 | 276.848 | 0.97
sage_ns | ogbn-products | 63.54111 | 52.0431 | 1.22
sage_ns | reddit | 66.257414 | 66.3105 | 1



</body>

</html>

The tests were conducted the following way:
dry run: 3 epochs
timing: 10 epochs
 
GAT
hidden feats: 120
num heads: 4
num layers: 3
dropout: 0.5
fanouts: 10, 10, 10
batch size: 6554
 
RGCN
hidden feats: 64
num bases: 2
num layers: 2
dropout: 0.5
self loop: True
fanouts: 25, 15
batch size: 1024
 
GraphSAGE

hidden feats: 256
num layers: 3
dropout: 0.5
fanouts: 5, 10, 15
batch size: 1000